Export the IPCAPS2 result to a text file

Export clustering result of ipcaps2 to text file called 'groups.txt'.

export.groups(result.dir, silence.mode = FALSE)

Arguments

result.dir
A result directory as the $output object returned from the ipcaps2 function.

silence.mode
To enable or disable silence mode. If silence mode is enabled, the fuction is processed without printing any message on the screen, and it is slightly faster. Default = TRUE.

Value

A data frame of exported data containing 4 columns; group, node, label, row.number, as described below for more details:

group represents group membership of IPCAPS2 result.

node represents node numbers of IPCAPS2 result.

label represents labels of rows in orginal input data.

row.number represents row numbers of orginal input data.

Details

After running, this function exports the file called 'groups.txt' to the same result directory. If 'groups.txt' already exists in the result directory, the exported file is changed to 'groups1.txt', 'groups2.txt', 'groups3.txt', ..., accordingly.

Examples

# Importantly, bed file, bim file, and fam file are required
# Use the example files embedded in the package

BED.file <- system.file("extdata","ipcaps_example.bed",package="IPCAPS2")
LABEL.file <- system.file("extdata","ipcaps_example_individuals.txt.gz",package="IPCAPS2")

my.cluster <- ipcaps2(bed=BED.file,label.file=LABEL.file,lab.col=2,out=tempdir(),silence=TRUE,no.rep=1)

#Here, to export the IPCAPS2 result to a text file
exported.data <- export.groups(my.cluster$output.dir)

#> Exporting node 3 as group 1
#> Exporting node 4 as group 2
#> Exporting node 5 as group 3
#> Exporting node 6 as group 4
#> Exporting node 7 as group 5
#> Note: save as /var/folders/sp/hhmj9xvx53z4g4dktf5f503r0000gp/T//RtmpY9IxUI/cluster_output/groups1.txt

print(dim(exported.data))

#> [1] 103   4

head(exported.data)

#>   group node    label row.number
#> 1     1    3 outlier3        101
#> 2     2    4 outlier3        102
#> 3     3    5 outlier3        103
#> 4     4    6     pop1          1
#> 5     4    6     pop1          2
#> 6     4    6     pop1          3