Workflow to generate HTML files for all kinds of plots

Generate HTML files and all image files (plots) from the result of ipcaps2. The clustering result is shown as a tree rendering by the online Google Organizational Chart library. Note that the Internet is required to view the HTML files.

save.plots(output.dir)

Arguments

output.dir
A result directory as the $output object returned from the ipcaps2 function.

Value

NULL

Details

After running, this function generates all plots and saves as image files in the sub-directory 'images'. It calls save.plots.cluster.html, save.eigenplots.html, and save.plots.label.html to generate all HTML files.

See also

Examples

# Importantly, bed file, bim file, and fam file are required
# Use the example files embedded in the package

BED.file <- system.file("extdata","ipcaps_example.bed",package="IPCAPS2")
LABEL.file <- system.file("extdata","ipcaps_example_individuals.txt.gz",package="IPCAPS2")

my.cluster <- ipcaps2(bed=BED.file,label.file=LABEL.file,lab.col=2,out=tempdir(),silence=TRUE,no.rep=1)

#Here, to generate all plots and HTML files
save.plots.label.html(my.cluster$output.dir)